refactor(rendering): modularize tile-related functions and improve imports
- Moved `getTile`, `getBiome`, and related functions from `utils.js` to a new `rendering/tile.js` module for better organization.
- Updated import paths across various files to reflect the new structure, enhancing code clarity and maintainability.
- Removed unused functions from `utils.js` to streamline the codebase.
